Add Support for Hubspace Lightbulbs
This adds support for Hubspace Lightbulbs. So far only tested on the BR30
Color CCT Light. This currently only supports setting of Temperature, but
color support will come soon. This also breaks supports for fan Lightbulbs
due to an incapability with the DeviceFunction enum with the attribute ID.
For no scaling values between the bulb and homekit are hardcoded and can be
read from values range received from the metadevice.device

This also fixes an issue with sending and receiving hex values that greater
than a single byte where the bytes were being send in reverse order.